Update shopping cart function

+++ modules/bookingmanagement/bookingmanagement.php
@@ -21,8 +21,14 @@ class BookingManagement extends Module
 			return false;
 		}
 
+		// Execute module install SQL statements
+		$sql_file = dirname(__FILE__).'/install/install.sql';
+		if (!$this->loadSQLFile($sql_file)) {
+			return false;
+		}
+
 		// Register hooks
-		if (!$this->registerHook('displayFooterProduct')) {
+		if (!$this->registerHook('displayFooterProduct') || !$this->registerHook('displayRightColumnProduct')) {
 			return false;
 		}
 
@@ -32,6 +38,7 @@ class BookingManagement extends Module
 		// All went well!
 		return true;
 	}
+
 	public function uninstall()
 	{
 		// Call uninstall parent method
@@ -39,11 +46,40 @@ class BookingManagement extends Module
 			return false;
 		}
 
+		// Execute module install SQL statements
+		$sql_file = dirname(__FILE__).'/install/uninstall.sql';
+		if (!$this->loadSQLFile($sql_file)) {
+			return false;
+		}
+
+
 		// Delete configuration values
 		Configuration::deleteByName('BOOKING_STANDARD_DAYS');
 
 		// All went well!
 		return true;
+	}
+
+	public function loadSQLFile($sql_file)
+	{
+		// Get install SQL file content
+		$sql_content = file_get_contents($sql_file);
+
+		// Replace prefix and store SQL command in array
+		$sql_content = str_replace('PREFIX_', _DB_PREFIX_, $sql_content);
+		$sql_requests = preg_split("/;\s*[\r\n]+/", $sql_content);
+
+		// Execute each SQL statement
+		$result = true;
+		foreach($sql_requests as $request)
+		{
+			if (!empty($request)) {
+				$result &= Db::getInstance()->execute(trim($request));
+			}
+		}
+
+		// Return result
+		return $result;
 	}
 
 	public function getHookController($hook_name)
@@ -71,5 +107,11 @@ class BookingManagement extends Module
 	{
 		$controller = $this->getHookController('displayFooterProduct');
 		return $controller->run();
+	}
+
+	public function hookDisplayRightColumnProduct()
+	{
+		$controller = $this->getHookController('displayRightColumnProduct');
+		return $controller->initContent();
 	}
 }
